small bugfis

+++ pycalcurse.py
@@ -691,7 +691,8 @@ class PyCalCurse(object):
                 self._refresh_after_popup()
                 return
             elif x in [ord('y'), ord('j')]:
-                os.remove(self.calendar_ressources[choosen_ressource].ressouce_path)
+                if self.calendar_ressources[choosen_ressource].ressource_type == 'webressource':
+                    os.remove(self.calendar_ressources[choosen_ressource].ressouce_path)
                 del self.calendar_ressources[choosen_ressource]
                 config_file_path = os.path.expanduser(
                     '~/.config/pycalcurse/ressources.csv'
